requirements.txt
streamlit
plotly
pandas
sqlalchemy
psycopg2-binary
streamlit_autorefresh
bcrypt


